feat(Controller): add 'PlayerManager' as a controller unit and apply to 'GameManager'
- extract 'Player' related parts from 'GameManager' to new class 'PlayerManager'
